# backend_api_base_netcore8/Application/DTOs/User/UserDto.cs
namespace backend_api_base_netcore8.Application.DTOs.User;

public class UserDto
{
    public int Id { get; set; }
    public string Username { get; set; } = string.Empty;
    public string FirstName { get; set; } = string.Empty;
    public string LastName { get; set; } = string.Empty;
    public string Email { get; set; } = string.Empty;
    public int? Phone { get; set; }
    public int RoleId { get; set; }
    public string Role { get; set; } = string.Empty;
}