# backend_api_base_netcore8/Application/DTOs/Users/CreateUserRequest.cs
namespace backend_api_base_netcore8.Application.DTOs.Users;

public class CreateUserRequest
{
    public string Username { get; set; } = string.Empty;
    public string FirstName { get; set; } = string.Empty;
    public string LastName { get; set; } = string.Empty;
    public string Email { get; set; } = string.Empty;
    public string Password { get; set; } = string.Empty;
    public int RoleId { get; set; }
    public int? Phone { get; set; }
}