# backend_api_base_netcore8/Application/DTOs/Users/UpdateUserRequest.cs
namespace backend_api_base_netcore8.Application.DTOs.Users;

public class UpdateUserRequest
{
    public int Id { get; set; }
    public string Username { get; set; } = string.Empty;
    public string FirstName { get; set; } = string.Empty;
    public string LastName { get; set; } = string.Empty;
    public string Email { get; set; } = string.Empty;
    public string? Password { get; set; }
    public int RoleId { get; set; }
    public int? Phone { get; set; }
}