# backend_api_base_netcore8/openapi.yaml
"openapi": "3.0.1"
"info":
  "title": "Auth API"
  "description": "API para autenticacion mediante JWT."
  "version": "v1"
"paths":
  "/api/auth/echoping":
    "get":
      "tags":
        - "Auth"
      "responses":
        "200":
          "description": "OK"
  "/api/auth/echouser":
    "get":
      "tags":
        - "Auth"
      "responses":
        "200":
          "description": "OK"
  "/api/auth/authenticate":
    "post":
      "tags":
        - "Auth"
      "summary": "Autentica un usuario y emite un JWT."
      "requestBody":
        "content":
          "application/json":
            "example":
              "username": "admin"
              "password": "P@ssw0rd!"
        "required": true
      "responses":
        "200":
          "description": "Login exitoso"
          "content":
            "application/json":
              "example":
                "token": "\u003cjwt\u003e"
                "expiresIn": 3600
        "400":
          "description": "Bad Request"
          "content":
            "text/plain":
              "schema":
                "$ref": "#/components/schemas/ApiErrorResponse"
            "application/json":
              "schema":
                "$ref": "#/components/schemas/ApiErrorResponse"
            "text/json":
              "schema":
                "$ref": "#/components/schemas/ApiErrorResponse"
        "401":
          "description": "Unauthorized"
          "content":
            "text/plain":
              "schema":
                "$ref": "#/components/schemas/ApiErrorResponse"
            "application/json":
              "schema":
                "$ref": "#/components/schemas/ApiErrorResponse"
            "text/json":
              "schema":
                "$ref": "#/components/schemas/ApiErrorResponse"
        "500":
          "description": "Internal Server Error"
          "content":
            "text/plain":
              "schema":
                "$ref": "#/components/schemas/ApiErrorResponse"
            "application/json":
              "schema":
                "$ref": "#/components/schemas/ApiErrorResponse"
            "text/json":
              "schema":
                "$ref": "#/components/schemas/ApiErrorResponse"
  "/api/auth/google":
    "post":
      "tags":
        - "Auth"
      "requestBody":
        "content":
          "application/json":
            "schema":
              "$ref": "#/components/schemas/GoogleLoginRequest"
          "text/json":
            "schema":
              "$ref": "#/components/schemas/GoogleLoginRequest"
          "application/*+json":
            "schema":
              "$ref": "#/components/schemas/GoogleLoginRequest"
      "responses":
        "200":
          "description": "OK"
          "content":
            "text/plain":
              "schema":
                "$ref": "#/components/schemas/LoginResponse"
            "application/json":
              "schema":
                "$ref": "#/components/schemas/LoginResponse"
            "text/json":
              "schema":
                "$ref": "#/components/schemas/LoginResponse"
        "400":
          "description": "Bad Request"
          "content":
            "text/plain":
              "schema":
                "$ref": "#/components/schemas/ApiErrorResponse"
            "application/json":
              "schema":
                "$ref": "#/components/schemas/ApiErrorResponse"
            "text/json":
              "schema":
                "$ref": "#/components/schemas/ApiErrorResponse"
        "401":
          "description": "Unauthorized"
          "content":
            "text/plain":
              "schema":
                "$ref": "#/components/schemas/ApiErrorResponse"
            "application/json":
              "schema":
                "$ref": "#/components/schemas/ApiErrorResponse"
            "text/json":
              "schema":
                "$ref": "#/components/schemas/ApiErrorResponse"
        "500":
          "description": "Internal Server Error"
          "content":
            "text/plain":
              "schema":
                "$ref": "#/components/schemas/ApiErrorResponse"
            "application/json":
              "schema":
                "$ref": "#/components/schemas/ApiErrorResponse"
            "text/json":
              "schema":
                "$ref": "#/components/schemas/ApiErrorResponse"
  "/api/auth/password":
    "post":
      "tags":
        - "Auth"
      "requestBody":
        "content":
          "application/json":
            "schema":
              "$ref": "#/components/schemas/GeneratePasswordRequest"
          "text/json":
            "schema":
              "$ref": "#/components/schemas/GeneratePasswordRequest"
          "application/*+json":
            "schema":
              "$ref": "#/components/schemas/GeneratePasswordRequest"
      "responses":
        "200":
          "description": "OK"
          "content":
            "text/plain":
              "schema":
                "$ref": "#/components/schemas/PasswordGenerationResponse"
            "application/json":
              "schema":
                "$ref": "#/components/schemas/PasswordGenerationResponse"
            "text/json":
              "schema":
                "$ref": "#/components/schemas/PasswordGenerationResponse"
        "404":
          "description": "Not Found"
          "content":
            "text/plain":
              "schema":
                "$ref": "#/components/schemas/ProblemDetails"
            "application/json":
              "schema":
                "$ref": "#/components/schemas/ProblemDetails"
            "text/json":
              "schema":
                "$ref": "#/components/schemas/ProblemDetails"
        "400":
          "description": "Bad Request"
          "content":
            "text/plain":
              "schema":
                "$ref": "#/components/schemas/ProblemDetails"
            "application/json":
              "schema":
                "$ref": "#/components/schemas/ProblemDetails"
            "text/json":
              "schema":
                "$ref": "#/components/schemas/ProblemDetails"
  "/api/users":
    "get":
      "tags":
        - "Users"
      "parameters":
        - "name": "page"
          "in": "query"
          "schema":
            "type": "integer"
            "format": "int32"
            "default": 1
        - "name": "perPage"
          "in": "query"
          "schema":
            "type": "integer"
            "format": "int32"
            "default": 20
      "responses":
        "200":
          "description": "OK"
          "content":
            "text/plain":
              "schema":
                "$ref": "#/components/schemas/UserResponseCollectionResponse"
            "application/json":
              "schema":
                "$ref": "#/components/schemas/UserResponseCollectionResponse"
            "text/json":
              "schema":
                "$ref": "#/components/schemas/UserResponseCollectionResponse"
        "400":
          "description": "Bad Request"
          "content":
            "text/plain":
              "schema":
                "$ref": "#/components/schemas/ApiErrorResponse"
            "application/json":
              "schema":
                "$ref": "#/components/schemas/ApiErrorResponse"
            "text/json":
              "schema":
                "$ref": "#/components/schemas/ApiErrorResponse"
        "401":
          "description": "Unauthorized"
          "content":
            "text/plain":
              "schema":
                "$ref": "#/components/schemas/ProblemDetails"
            "application/json":
              "schema":
                "$ref": "#/components/schemas/ProblemDetails"
            "text/json":
              "schema":
                "$ref": "#/components/schemas/ProblemDetails"
        "403":
          "description": "Forbidden"
          "content":
            "text/plain":
              "schema":
                "$ref": "#/components/schemas/ProblemDetails"
            "application/json":
              "schema":
                "$ref": "#/components/schemas/ProblemDetails"
            "text/json":
              "schema":
                "$ref": "#/components/schemas/ProblemDetails"
      "security":
        - "Bearer":
    "post":
      "tags":
        - "Users"
      "requestBody":
        "content":
          "application/json":
            "schema":
              "$ref": "#/components/schemas/CreateUserRequest"
          "text/json":
            "schema":
              "$ref": "#/components/schemas/CreateUserRequest"
          "application/*+json":
            "schema":
              "$ref": "#/components/schemas/CreateUserRequest"
      "responses":
        "200":
          "description": "OK"
          "content":
            "text/plain":
              "schema":
                "type": "string"
            "application/json":
              "schema":
                "type": "string"
            "text/json":
              "schema":
                "type": "string"
        "401":
          "description": "Unauthorized"
        "403":
          "description": "Forbidden"
      "security":
        - "Bearer":
    "put":
      "tags":
        - "Users"
      "requestBody":
        "content":
          "application/json":
            "schema":
              "$ref": "#/components/schemas/UpdateUserRequest"
          "text/json":
            "schema":
              "$ref": "#/components/schemas/UpdateUserRequest"
          "application/*+json":
            "schema":
              "$ref": "#/components/schemas/UpdateUserRequest"
      "responses":
        "200":
          "description": "OK"
          "content":
            "text/plain":
              "schema":
                "type": "string"
            "application/json":
              "schema":
                "type": "string"
            "text/json":
              "schema":
                "type": "string"
        "404":
          "description": "Not Found"
          "content":
            "text/plain":
              "schema":
                "$ref": "#/components/schemas/ProblemDetails"
            "application/json":
              "schema":
                "$ref": "#/components/schemas/ProblemDetails"
            "text/json":
              "schema":
                "$ref": "#/components/schemas/ProblemDetails"
        "401":
          "description": "Unauthorized"
        "403":
          "description": "Forbidden"
      "security":
        - "Bearer":
  "/api/users/{id}":
    "get":
      "tags":
        - "Users"
      "parameters":
        - "name": "id"
          "in": "path"
          "required": true
          "schema":
            "type": "integer"
            "format": "int32"
      "responses":
        "200":
          "description": "OK"
          "content":
            "text/plain":
              "schema":
                "$ref": "#/components/schemas/UserResponse"
            "application/json":
              "schema":
                "$ref": "#/components/schemas/UserResponse"
            "text/json":
              "schema":
                "$ref": "#/components/schemas/UserResponse"
        "404":
          "description": "Not Found"
          "content":
            "text/plain":
              "schema":
                "$ref": "#/components/schemas/ProblemDetails"
            "application/json":
              "schema":
                "$ref": "#/components/schemas/ProblemDetails"
            "text/json":
              "schema":
                "$ref": "#/components/schemas/ProblemDetails"
        "401":
          "description": "Unauthorized"
        "403":
          "description": "Forbidden"
      "security":
        - "Bearer":
    "delete":
      "tags":
        - "Users"
      "parameters":
        - "name": "id"
          "in": "path"
          "required": true
          "schema":
            "type": "integer"
            "format": "int32"
      "responses":
        "200":
          "description": "OK"
          "content":
            "text/plain":
              "schema":
                "type": "string"
            "application/json":
              "schema":
                "type": "string"
            "text/json":
              "schema":
                "type": "string"
        "404":
          "description": "Not Found"
          "content":
            "text/plain":
              "schema":
                "$ref": "#/components/schemas/ProblemDetails"
            "application/json":
              "schema":
                "$ref": "#/components/schemas/ProblemDetails"
            "text/json":
              "schema":
                "$ref": "#/components/schemas/ProblemDetails"
        "401":
          "description": "Unauthorized"
        "403":
          "description": "Forbidden"
      "security":
        - "Bearer":
  "/api/users/roles":
    "get":
      "tags":
        - "Users"
      "responses":
        "200":
          "description": "OK"
          "content":
            "text/plain":
              "schema":
                "type": "array"
                "items":
                  "$ref": "#/components/schemas/RoleResponse"
            "application/json":
              "schema":
                "type": "array"
                "items":
                  "$ref": "#/components/schemas/RoleResponse"
            "text/json":
              "schema":
                "type": "array"
                "items":
                  "$ref": "#/components/schemas/RoleResponse"
        "401":
          "description": "Unauthorized"
        "403":
          "description": "Forbidden"
      "security":
        - "Bearer":
"components":
  "schemas":
    "ApiErrorResponse":
      "type": "object"
      "properties":
        "success":
          "type": "boolean"
        "messageError":
          "type": "string"
          "nullable": true
      "additionalProperties": false
    "CollectionMeta":
      "type": "object"
      "properties":
        "page":
          "type": "integer"
          "format": "int32"
        "perPage":
          "type": "integer"
          "format": "int32"
        "total":
          "type": "integer"
          "format": "int32"
        "totalPages":
          "type": "integer"
          "format": "int32"
      "additionalProperties": false
    "CreateUserRequest":
      "type": "object"
      "properties":
        "username":
          "type": "string"
          "nullable": true
        "firstName":
          "type": "string"
          "nullable": true
        "lastName":
          "type": "string"
          "nullable": true
        "email":
          "type": "string"
          "nullable": true
        "password":
          "type": "string"
          "nullable": true
        "roleId":
          "type": "integer"
          "format": "int32"
        "phone":
          "type": "integer"
          "format": "int32"
          "nullable": true
      "additionalProperties": false
    "GeneratePasswordRequest":
      "type": "object"
      "properties":
        "userId":
          "type": "integer"
          "format": "int32"
        "length":
          "type": "integer"
          "format": "int32"
      "additionalProperties": false
    "GoogleLoginRequest":
      "type": "object"
      "properties":
        "idToken":
          "type": "string"
          "nullable": true
      "additionalProperties": false
    "LoginRequest":
      "type": "object"
      "properties":
        "email":
          "type": "string"
          "nullable": true
        "password":
          "type": "string"
          "nullable": true
      "additionalProperties": false
    "LoginResponse":
      "type": "object"
      "properties":
        "userId":
          "type": "integer"
          "format": "int32"
        "status":
          "type": "integer"
          "format": "int32"
        "token":
          "type": "string"
          "nullable": true
        "expiresIn":
          "type": "integer"
          "format": "int32"
        "user":
          "$ref": "#/components/schemas/UserDto"
      "additionalProperties": false
    "PasswordGenerationResponse":
      "type": "object"
      "properties":
        "userId":
          "type": "integer"
          "format": "int32"
        "password":
          "type": "string"
          "nullable": true
      "additionalProperties": false
    "ProblemDetails":
      "type": "object"
      "properties":
        "type":
          "type": "string"
          "nullable": true
        "title":
          "type": "string"
          "nullable": true
        "status":
          "type": "integer"
          "format": "int32"
          "nullable": true
        "detail":
          "type": "string"
          "nullable": true
        "instance":
          "type": "string"
          "nullable": true
      "additionalProperties":
    "RoleResponse":
      "type": "object"
      "properties":
        "id":
          "type": "integer"
          "format": "int32"
        "name":
          "type": "string"
          "nullable": true
      "additionalProperties": false
    "UpdateUserRequest":
      "type": "object"
      "properties":
        "id":
          "type": "integer"
          "format": "int32"
        "username":
          "type": "string"
          "nullable": true
        "firstName":
          "type": "string"
          "nullable": true
        "lastName":
          "type": "string"
          "nullable": true
        "email":
          "type": "string"
          "nullable": true
        "password":
          "type": "string"
          "nullable": true
        "roleId":
          "type": "integer"
          "format": "int32"
        "phone":
          "type": "integer"
          "format": "int32"
          "nullable": true
      "additionalProperties": false
    "UserDto":
      "type": "object"
      "properties":
        "id":
          "type": "integer"
          "format": "int32"
        "username":
          "type": "string"
          "nullable": true
        "firstName":
          "type": "string"
          "nullable": true
        "lastName":
          "type": "string"
          "nullable": true
        "email":
          "type": "string"
          "nullable": true
        "phone":
          "type": "integer"
          "format": "int32"
          "nullable": true
        "roleId":
          "type": "integer"
          "format": "int32"
        "role":
          "type": "string"
          "nullable": true
      "additionalProperties": false
    "UserResponse":
      "type": "object"
      "properties":
        "id":
          "type": "integer"
          "format": "int32"
        "username":
          "type": "string"
          "nullable": true
        "firstName":
          "type": "string"
          "nullable": true
        "lastName":
          "type": "string"
          "nullable": true
        "email":
          "type": "string"
          "nullable": true
        "roleId":
          "type": "integer"
          "format": "int32"
        "role":
          "type": "string"
          "nullable": true
        "phone":
          "type": "integer"
          "format": "int32"
          "nullable": true
      "additionalProperties": false
    "UserResponseCollectionResponse":
      "type": "object"
      "properties":
        "success":
          "type": "boolean"
        "data":
          "type": "array"
          "items":
            "$ref": "#/components/schemas/UserResponse"
          "nullable": true
        "meta":
          "$ref": "#/components/schemas/CollectionMeta"
      "additionalProperties": false
  "securitySchemes":
    "Bearer":
      "type": "http"
      "description": "Introduce el token JWT con el esquema Bearer."
      "scheme": "bearer"
      "bearerFormat": "JWT"